Removed git of ble-module

+++ imu_general/utilities/i2c_sp.h
@@ -16,7 +16,7 @@ extern "C" {
 #include <stdint.h>
 
 #ifdef NONRASPI
-#include "../tests/test_stubs.h"
+#include "../tests/nonpi_stubs.h"
 #else
 #include <linux/i2c-dev.h>
 #include <sys/ioctl.h>
@@ -24,8 +24,8 @@ extern "C" {
 #include <unistd.h>
 #endif
 
-extern size_t file;
-
+extern int file;
+    
 bool I2C_Init( void );
 bool I2C_Deinit( void );
 bool I2C_Read( uint8_t addr, uint8_t *i2c_read_data, uint8_t i2c_read_data_length);
@@ -33,8 +33,9 @@ uint8_t I2C_Read_Reg( uint8_t addr, uint8_t reg);
 bool I2C_Read_Regs( uint8_t addr, uint8_t reg, uint8_t *i2c_read_data, uint8_t i2c_read_data_length);
 bool I2C_Write( uint8_t addr, uint8_t *i2c_write_data, uint8_t i2c_write_data_length);
 
+    
 #ifdef __cplusplus
 }
 #endif
-    
+
 #endif /* SYSTEM_I2C_SP_H_ */
